+++ app/controllers/carts_controller.rb
@@ -2,7 +2,7 @@
 
 class CartsController < ApplicationController
   def create
-    session[:project_item_item] = params[:project_item]
+    session[:project_item] = params[:project_item]
     redirect_to new_project_item_path
   end
 end