+++ db/migrate/20181012081942_add_iid_to_order_change.rb
@@ -1,5 +1,0 @@
-class AddIidToOrderChange < ActiveRecord::Migration[5.2]
-  def change
-    add_column :order_changes, :iid, :integer, index: true
-  end
-end