+++ db/migrate/20181012081942_add_iid_to_project_item_change.rb
@@ -1,0 +1,5 @@
+class AddIidToProjectItemChange < ActiveRecord::Migration[5.2]
+  def change
+    add_column :project_item_changes, :iid, :integer, index: true
+  end
+end