+++ spec/models/project_item_change_spec.rb
@@ -2,7 +2,7 @@
 
 require "rails_helper"
 
-RSpec.describe OrderChange, type: :model do
+RSpec.describe ProjectItemChange, type: :model do
   it { should belong_to(:project_item) }
   it { should belong_to(:author) }
 
@@ -10,25 +10,25 @@ RSpec.describe OrderChange, type: :model do
     it "is true when project_item change is created by project_item owner" do
       owner = create(:user)
       project_item = create(:project_item, user: owner)
-      order_change = create(:order_change, project_item: project_item, author: owner,
+      project_item_change = create(:project_item_change, project_item: project_item, author: owner,
                             message: "some question")
 
-      expect(order_change).to be_question
+      expect(project_item_change).to be_question
     end
 
     it "is false when there is not change author" do
-      order_change = create(:order_change, author: nil)
+      project_item_change = create(:project_item_change, author: nil)
 
-      expect(order_change).to_not be_question
+      expect(project_item_change).to_not be_question
     end
 
     it "is false when change author is not project_item owner" do
       owner = create(:user)
       project_item = create(:project_item, user: owner)
-      order_change = create(:order_change, project_item: project_item, author: create(:user),
+      project_item_change = create(:project_item_change, project_item: project_item, author: create(:user),
                             message: "some question")
 
-      expect(order_change).to_not be_question
+      expect(project_item_change).to_not be_question
     end
   end
 end