+++ CHANGELOG.md
@@ -73,7 +73,8 @@ Please view this file on the master branch, on stable branches it's out of date.
 - Update services and category list in data.yml (@goreck888)
 - Update home view with static version (@kmarszalek)
 - Upgrade service detail view (@kmarszalek)
-- Rename `Order` into `ProjectItem` (@mkasztelnik)
+- Rename `Order` to `ProjectItem` (@mkasztelnik)
+- Rename `OrderChange` to `ProjectItemChange` (@mkasztelnik)
 
 ### Deprecated